requirements.txt
streamlit
extra-streamlit-components
google-genai
streamlit-paste-button
gpt4all
requests
pillow
huggingface_hub